"eviter des warning sur unlink d'un repertoire (francois schreuer)"
git-svn-id: svn://trac.rezo.net/spip/spip@9905 caf5f3e8-d4fe-0310-bb3e-c32d5e47d55d

+++ ecrire/inc/flock.php
@@ -162,11 +162,14 @@ function supprimer_fichier($fichier, $lock=true) {
 	}
 	
 	// supprimer
-	unlink($fichier);
+	@unlink($fichier);
 }
 // Supprimer brutalement, si le fichier existe
 function spip_unlink($fichier) {
-	supprimer_fichier($fichier,false);
+	if (is_dir($fichier))
+		@rmdir($fichier);
+	else 
+		supprimer_fichier($fichier,false);
 }
 
 //
